+++ handleAll.h
@@ -31,7 +31,7 @@ bool recvAll(SOCKET s, char* p, size_t len, const std::string& filename);
 bool recvAll(SOCKET s, char* p, size_t len, const std::string& filename, const std::string& account);
 bool recvevery(SOCKET s, char* p, size_t len, int flag);
 bool sendAll(SOCKET s, char* p, size_t len, FILE* stream);
-bool sendevery(SOCKET s, const char* p, int len, int flag);
+bool sendevery(SOCKET s, const char* p, size_t len, int flag);
 std::string getfilename(const std::string& str);
 bool deleteFile(const std::string& filename);
